Data is now fetched from local storage instead of the database when profile page refreshes. When employee logs in, employeeData is fetched from the database only once.
