+++ src/app/app.component.ts
@@ -6,12 +6,18 @@ import { RouterModule } from '@angular/router';
   imports: [RouterModule],
   selector: 'app-root',
   templateUrl: './app.component.html',
-  styleUrl: './app.component.css',
+  styleUrls: ['./app.component.css'],
   encapsulation: ViewEncapsulation.Emulated
 })
 export class AppComponent {
   title = 'Events-System';
   isDrawerThin = false;
+  employeeData: any; // Define employeeData property
+
+  constructor() {
+    // Initialize employeeData from localStorage or any other source
+    this.employeeData = JSON.parse(localStorage.getItem('employeeData') || '{}');
+  }
 
   toggleDrawer() {
     this.isDrawerThin = !this.isDrawerThin;
